refactor(workspace): extract useWorkspaceLayout hook
Move chat panel width and divider drag handlers into a dedicated hook.
