+++ frontend/src/pages/Workspace/index.tsx
@@ -88,6 +88,7 @@ import { PendingRerunBanner } from './PendingRerunBanner';
 import { ProjectDetailsDialog } from './ProjectDetailsDialog';
 import { SpreadsheetChrome } from './SpreadsheetChrome';
 import { useAddDocuments } from './hooks/useAddDocuments';
+import { useWorkspaceLayout } from './hooks/useWorkspaceLayout';
 import { SpreadsheetSurface } from './SpreadsheetSurface';
 import type {
   CellFormatMap,
@@ -183,11 +184,12 @@ function Workspace() {
   const [sourceDocReloadToken, setSourceDocReloadToken] = useState(0);
   const [attachingSourceDocs, setAttachingSourceDocs] = useState(false);
   const sourceDocInputRef = useRef<HTMLInputElement | null>(null);
-  const [chatWidth, setChatWidth] = useState(() => {
-    const saved = Number(localStorage.getItem('workspace.chatWidth'));
-    return Number.isFinite(saved) ? saved : 380;
-  });
-  const [isDraggingDivider, setIsDraggingDivider] = useState(false);
+  const {
+    chatWidth,
+    setChatWidth,
+    isDraggingDivider,
+    startDividerDrag,
+  } = useWorkspaceLayout();
   const [reextraction, setReextraction] = useState<WorkspaceReextractionState | null>(null);
   const [wsConnected, setWsConnected] = useState(false);
   const [reextractConfirm, setReextractConfirm] = useState<{ columns: string[] } | null>(null);
@@ -1050,29 +1052,6 @@ function Workspace() {
     ? `Re-extracting ${reextraction.columns.join(', ')} (${reextraction.processedDocuments}/${reextraction.totalDocuments || '?'} docs)`
     : (status?.current_step || status?.status || 'No project status');
   const bottombarProgress = reextraction ? reextractionPercent : progressPercent;
-
-  const startDividerDrag = useCallback((event: React.PointerEvent<HTMLDivElement>) => {
-    event.preventDefault();
-    setIsDraggingDivider(true);
-
-    const handleMove = (moveEvent: PointerEvent) => {
-      const nextWidth = window.innerWidth - moveEvent.clientX;
-      const maxWidth = Math.max(24, window.innerWidth - 24);
-      const clamped = Math.min(maxWidth, Math.max(0, nextWidth));
-      const snapped = clamped < 56 ? 0 : clamped > window.innerWidth - 80 ? window.innerWidth : clamped;
-      setChatWidth(snapped);
-      localStorage.setItem('workspace.chatWidth', String(snapped));
-    };
-
-    const handleUp = () => {
-      setIsDraggingDivider(false);
-      window.removeEventListener('pointermove', handleMove);
-      window.removeEventListener('pointerup', handleUp);
-    };
-
-    window.addEventListener('pointermove', handleMove);
-    window.addEventListener('pointerup', handleUp);
-  }, []);
 
   // Source document of the currently selected data row, used by the optional
   // source panel on the Data sheet. Uses the raw _source_document value so it
